Android-Refactor: Remove unused methods and classes from color utilities
This commit refactors the color utilities by removing unused methods and classes, including:

- Removed `StringUtils` class.
- Removed several methods from `ColorUtils`, such as `alphaFromArgb`, `isOpaque`, `argbFromXyz`, `labFromArgb`, `argbFromLab`, `lstarFromY`.
- Removed `rotationDirection` and `differenceDegrees` methods from `MathUtils`.
- Removed unused methods and variables from `Cam16`.
- Removed `inViewingConditions` method from `Hct`.

These changes streamline the codebase and remove unnecessary functionality.
