+++ Android/mobile/src/main/java/io/material/hct/material/hct/Cam16.kt
@@ -13,16 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.material.hct
+package io.material.hct.material.hct
 
-import io.material.utils.ColorUtils
+import io.material.hct.material.utils.ColorUtils
+import kotlin.math.PI
 import kotlin.math.abs
 import kotlin.math.atan2
 import kotlin.math.cos
-import kotlin.math.expm1
 import kotlin.math.hypot
 import kotlin.math.ln1p
-import kotlin.math.max
 import kotlin.math.pow
 import kotlin.math.sign
 import kotlin.math.sin
@@ -63,7 +62,8 @@ class Cam16
  * @param jstar CAM16-UCS J coordinate
  * @param astar CAM16-UCS a coordinate
  * @param bstar CAM16-UCS b coordinate
- */ private constructor(
+ */
+private constructor(
     /** Hue in CAM16  */
     // CAM16 color dimensions, see getters for documentation.
     val hue: Double,
@@ -105,101 +105,12 @@ class Cam16
     val bstar: Double
 ) {
 
-
-    // Avoid allocations during conversion by pre-allocating an array.
-    private val tempArray = doubleArrayOf(0.0, 0.0, 0.0)
-
-    /**
-     * CAM16 instances also have coordinates in the CAM16-UCS space, called J*, a*, b*, or jstar,
-     * astar, bstar in code. CAM16-UCS is included in the CAM16 specification, and is used to measure
-     * distances between colors.
-     */
-    fun distance(other: Cam16): Double {
-        val dJ = jstar - other.jstar
-        val dA = astar - other.astar
-        val dB = bstar - other.bstar
-        val dEPrime = sqrt(dJ * dJ + dA * dA + dB * dB)
-        return 1.41 * dEPrime.pow(0.63)
-    }
-
-    /**
-     * ARGB representation of the color. Assumes the color was viewed in default viewing conditions,
-     * which are near-identical to the default viewing conditions for sRGB.
-     */
-    fun toInt(): Int {
-        return viewed(ViewingConditions.Companion.DEFAULT)
-    }
-
-    /**
-     * ARGB representation of the color, in defined viewing conditions.
-     *
-     * @param viewingConditions Information about the environment where the color will be viewed.
-     * @return ARGB representation of color
-     */
-    fun viewed(viewingConditions: ViewingConditions): Int {
-        val xyz = xyzInViewingConditions(viewingConditions, tempArray)
-        return ColorUtils.argbFromXyz(xyz[0], xyz[1], xyz[2])
-    }
-
-    fun xyzInViewingConditions(
-        viewingConditions: ViewingConditions,
-        returnArray: DoubleArray?
-    ): DoubleArray {
-        val alpha = if (chroma == 0.0 || j == 0.0) 0.0 else chroma / sqrt(
-            j / 100.0
-        )
-        val t = (alpha / (1.64 - 0.29.pow(viewingConditions.n)).pow(0.73)).pow(1.0 / 0.9)
-        val hRad = hue.toRadians()
-
-        val eHue = 0.25 * (cos(hRad + 2.0) + 3.8)
-        val ac = (viewingConditions.aw
-                * (j / 100.0).pow(1.0 / viewingConditions.c / viewingConditions.z))
-        val p1 = eHue * (50000.0 / 13.0) * viewingConditions.nc * viewingConditions.ncb
-        val p2 = ac / viewingConditions.nbb
-        val hSin = sin(hRad)
-        val hCos = cos(hRad)
-        val gamma = 23.0 * (p2 + 0.305) * t / (23.0 * p1 + 11.0 * t * hCos + 108.0 * t * hSin)
-        val a = gamma * hCos
-        val b = gamma * hSin
-        val rA = (460.0 * p2 + 451.0 * a + 288.0 * b) / 1403.0
-        val gA = (460.0 * p2 - 891.0 * a - 261.0 * b) / 1403.0
-        val bA = (460.0 * p2 - 220.0 * a - 6300.0 * b) / 1403.0
-        val rCBase = max(0.0, 27.13 * abs(rA) / (400.0 - abs(rA)))
-        val rC = sign(rA) * (100.0 / viewingConditions.fl) * rCBase.pow(1.0 / 0.42)
-        val gCBase = max(0.0, 27.13 * abs(gA) / (400.0 - abs(gA)))
-        val gC = sign(gA) * (100.0 / viewingConditions.fl) * gCBase.pow(1.0 / 0.42)
-        val bCBase = max(0.0, 27.13 * abs(bA) / (400.0 - abs(bA)))
-        val bC = sign(bA) * (100.0 / viewingConditions.fl) * bCBase.pow(1.0 / 0.42)
-        val rF = rC / viewingConditions.rgbD[0]
-        val gF = gC / viewingConditions.rgbD[1]
-        val bF = bC / viewingConditions.rgbD[2]
-        val matrix = CAM16RGB_TO_XYZ
-        val x = rF * matrix[0][0] + gF * matrix[0][1] + bF * matrix[0][2]
-        val y = rF * matrix[1][0] + gF * matrix[1][1] + bF * matrix[1][2]
-        val z = rF * matrix[2][0] + gF * matrix[2][1] + bF * matrix[2][2]
-        return if (returnArray != null) {
-            returnArray[0] = x
-            returnArray[1] = y
-            returnArray[2] = z
-            returnArray
-        } else {
-            doubleArrayOf(x, y, z)
-        }
-    }
-
     companion object {
         // Transforms XYZ color space coordinates to 'cone'/'RGB' responses in CAM16.
         val XYZ_TO_CAM16RGB = arrayOf(
             doubleArrayOf(0.401288, 0.650173, -0.051461),
             doubleArrayOf(-0.250268, 1.204414, 0.045854),
             doubleArrayOf(-0.002079, 0.048952, 0.953127)
-        )
-
-        // Transforms 'cone'/'RGB' responses in CAM16 to XYZ color space coordinates.
-        val CAM16RGB_TO_XYZ = arrayOf(
-            doubleArrayOf(1.8620678, -1.0112547, 0.14918678),
-            doubleArrayOf(0.38752654, 0.62144744, -0.00897398),
-            doubleArrayOf(-0.01584150, -0.03412294, 1.0499644)
         )
 
         /**
@@ -303,83 +214,8 @@ class Cam16
             return Cam16(hue, c, j, q, m, s, jstar, astar, bstar)
         }
 
-        /**
-         * @param j CAM16 lightness
-         * @param c CAM16 chroma
-         * @param h CAM16 hue
-         */
-        fun fromJch(j: Double, c: Double, h: Double): Cam16 {
-            return fromJchInViewingConditions(j, c, h, ViewingConditions.Companion.DEFAULT)
-        }
+        private fun Double.toRadians() = this * PI / 180.0
 
-        /**
-         * @param j CAM16 lightness
-         * @param c CAM16 chroma
-         * @param h CAM16 hue
-         * @param viewingConditions Information about the environment where the color was observed.
-         */
-        private fun fromJchInViewingConditions(
-            j: Double, c: Double, h: Double, viewingConditions: ViewingConditions
-        ): Cam16 {
-            val q = ((4.0
-                    / viewingConditions.c) * sqrt(j / 100.0)
-                    * (viewingConditions.aw + 4.0)
-                    * viewingConditions.flRoot)
-            val m = c * viewingConditions.flRoot
-            val alpha = c / sqrt(j / 100.0)
-            val s = 50.0 * sqrt(alpha * viewingConditions.c / (viewingConditions.aw + 4.0))
-            val hueRadians = h.toRadians()
-            val jstar = (1.0 + 100.0 * 0.007) * j / (1.0 + 0.007 * j)
-            val mstar = 1.0 / 0.0228 * ln1p(0.0228 * m)
-            val astar = mstar * cos(hueRadians)
-            val bstar = mstar * sin(hueRadians)
-            return Cam16(h, c, j, q, m, s, jstar, astar, bstar)
-        }
-
-        /**
-         * Create a CAM16 color from CAM16-UCS coordinates.
-         *
-         * @param jstar CAM16-UCS lightness.
-         * @param astar CAM16-UCS a dimension. Like a* in L*a*b*, it is a Cartesian coordinate on the Y
-         * axis.
-         * @param bstar CAM16-UCS b dimension. Like a* in L*a*b*, it is a Cartesian coordinate on the X
-         * axis.
-         */
-        fun fromUcs(jstar: Double, astar: Double, bstar: Double): Cam16 {
-            return fromUcsInViewingConditions(
-                jstar,
-                astar,
-                bstar,
-                ViewingConditions.Companion.DEFAULT
-            )
-        }
-
-        /**
-         * Create a CAM16 color from CAM16-UCS coordinates in defined viewing conditions.
-         *
-         * @param jstar CAM16-UCS lightness.
-         * @param astar CAM16-UCS a dimension. Like a* in L*a*b*, it is a Cartesian coordinate on the Y
-         * axis.
-         * @param bstar CAM16-UCS b dimension. Like a* in L*a*b*, it is a Cartesian coordinate on the X
-         * axis.
-         * @param viewingConditions Information about the environment where the color was observed.
-         */
-        fun fromUcsInViewingConditions(
-            jstar: Double, astar: Double, bstar: Double, viewingConditions: ViewingConditions
-        ): Cam16 {
-            val m = hypot(astar, bstar)
-            val m2 = expm1(m * 0.0228) / 0.0228
-            val c = m2 / viewingConditions.flRoot
-            var h = atan2(bstar, astar) * (180.0 / kotlin.math.PI)
-            if (h < 0.0) {
-                h += 360.0
-            }
-            val j = jstar / (1.0 - (jstar - 100.0) * 0.007)
-            return fromJchInViewingConditions(j, c, h, viewingConditions)
-        }
-
-        private inline fun Double.toRadians() = this * kotlin.math.PI / 180.0
-
-        private inline fun Double.toDegrees() = this * 180.0 / kotlin.math.PI
+        private fun Double.toDegrees() = this * 180.0 / PI
     }
 }
