+++ Android/mobile/src/main/java/io/material/hct/material/hct/HctSolver.kt
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-// This file is automatically generated. Do not modify it.
-package io.material.hct
+package io.material.hct.material.hct
 
-import io.material.utils.ColorUtils
-import io.material.utils.MathUtils
+import io.material.hct.material.utils.ColorUtils
+import io.material.hct.material.utils.MathUtils
+import kotlin.math.PI
 import kotlin.math.abs
 import kotlin.math.atan2
 import kotlin.math.ceil
@@ -314,7 +314,7 @@ object HctSolver {
      * @return A coterminal angle between 0 and 2pi.
      */
     fun sanitizeRadians(angle: Double): Double {
-        return (angle + kotlin.math.PI * 8) % (kotlin.math.PI * 2)
+        return (angle + PI * 8) % (PI * 2)
     }
 
     /**
@@ -325,8 +325,7 @@ object HctSolver {
      */
     fun trueDelinearized(rgbComponent: Double): Double {
         val normalized = rgbComponent / 100.0
-        var delinearized = 0.0
-        delinearized = if (normalized <= 0.0031308) {
+        val delinearized: Double = if (normalized <= 0.0031308) {
             normalized * 12.92
         } else {
             1.055 * normalized.pow(1.0 / 2.4) - 0.055
@@ -517,8 +516,8 @@ object HctSolver {
         var right = segment[1]
         for (axis in 0..2) {
             if (left[axis] != right[axis]) {
-                var lPlane = -1
-                var rPlane = 255
+                var lPlane: Int
+                var rPlane: Int
                 if (left[axis] < right[axis]) {
                     lPlane = criticalPlaneBelow(
                         trueDelinearized(
@@ -655,7 +654,7 @@ object HctSolver {
             return ColorUtils.argbFromLstar(lstar)
         }
         hueDegrees = MathUtils.sanitizeDegreesDouble(hueDegrees)
-        val hueRadians = hueDegrees / 180 * kotlin.math.PI
+        val hueRadians = hueDegrees / 180 * PI
         val y = ColorUtils.yFromLstar(lstar)
         val exactAnswer = findResultByJ(hueRadians, chroma, y)
         if (exactAnswer != 0) {
@@ -663,19 +662,5 @@ object HctSolver {
         }
         val linrgb = bisectToLimit(y, hueRadians)
         return ColorUtils.argbFromLinrgb(linrgb)
-    }
-
-    /**
-     * Finds an sRGB color with the given hue, chroma, and L*, if possible.
-     *
-     * @param hueDegrees The desired hue, in degrees.
-     * @param chroma The desired chroma.
-     * @param lstar The desired L*.
-     * @return An CAM16 object representing the sRGB color. The color has sufficiently close hue,
-     * chroma, and L* to the desired values, if possible; otherwise, the hue and L* will be
-     * sufficiently close, and chroma will be maximized.
-     */
-    fun solveToCam(hueDegrees: Double, chroma: Double, lstar: Double): Cam16 {
-        return Cam16.Companion.fromInt(solveToInt(hueDegrees, chroma, lstar))
     }
 }
