+++ Android/mobile/src/main/java/io/material/hct/material/utils/ColorUtils.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 // This file is automatically generated. Do not modify it.
-package io.material.utils
+package io.material.hct.material.utils
 
 import kotlin.math.pow
 import kotlin.math.roundToLong
@@ -32,15 +32,6 @@ object ColorUtils {
         doubleArrayOf(0.2126, 0.7152, 0.0722),
         doubleArrayOf(0.01932141, 0.11916382, 0.95034478)
     )
-    val XYZ_TO_SRGB = arrayOf(
-        doubleArrayOf(
-            3.2413774792388685, -1.5376652402851851, -0.49885366846268053
-        ), doubleArrayOf(
-            -0.9691452513005321, 1.8758853451067872, 0.04156585616912061
-        ), doubleArrayOf(
-            0.05562093689691305, -0.20395524564742123, 1.0571799111220335
-        )
-    )
     val WHITE_POINT_D65 = doubleArrayOf(95.047, 100.0, 108.883)
 
     /** Converts a color from RGB components to ARGB format.  */
@@ -54,11 +45,6 @@ object ColorUtils {
         val g = delinearized(linrgb[1])
         val b = delinearized(linrgb[2])
         return argbFromRgb(r, g, b)
-    }
-
-    /** Returns the alpha component of a color in ARGB format.  */
-    fun alphaFromArgb(argb: Int): Int {
-        return argb shr 24 and 255
     }
 
     /** Returns the red component of a color in ARGB format.  */
@@ -76,71 +62,12 @@ object ColorUtils {
         return argb and 255
     }
 
-    /** Returns whether a color in ARGB format is opaque.  */
-    fun isOpaque(argb: Int): Boolean {
-        return alphaFromArgb(argb) >= 255
-    }
-
-    /** Converts a color from ARGB to XYZ.  */
-    fun argbFromXyz(x: Double, y: Double, z: Double): Int {
-        val matrix = XYZ_TO_SRGB
-        val linearR = matrix[0][0] * x + matrix[0][1] * y + matrix[0][2] * z
-        val linearG = matrix[1][0] * x + matrix[1][1] * y + matrix[1][2] * z
-        val linearB = matrix[2][0] * x + matrix[2][1] * y + matrix[2][2] * z
-        val r = delinearized(linearR)
-        val g = delinearized(linearG)
-        val b = delinearized(linearB)
-        return argbFromRgb(r, g, b)
-    }
-
     /** Converts a color from XYZ to ARGB.  */
     fun xyzFromArgb(argb: Int): DoubleArray? {
         val r = linearized(redFromArgb(argb))
         val g = linearized(greenFromArgb(argb))
         val b = linearized(blueFromArgb(argb))
         return MathUtils.matrixMultiply(doubleArrayOf(r, g, b), SRGB_TO_XYZ)
-    }
-
-    /** Converts a color represented in Lab color space into an ARGB integer.  */
-    fun argbFromLab(l: Double, a: Double, b: Double): Int {
-        val whitePoint = WHITE_POINT_D65
-        val fy = (l + 16.0) / 116.0
-        val fx = a / 500.0 + fy
-        val fz = fy - b / 200.0
-        val xNormalized = labInvf(fx)
-        val yNormalized = labInvf(fy)
-        val zNormalized = labInvf(fz)
-        val x = xNormalized * whitePoint[0]
-        val y = yNormalized * whitePoint[1]
-        val z = zNormalized * whitePoint[2]
-        return argbFromXyz(x, y, z)
-    }
-
-    /**
-     * Converts a color from ARGB representation to L*a*b* representation.
-     *
-     * @param argb the ARGB representation of a color
-     * @return a Lab object representing the color
-     */
-    fun labFromArgb(argb: Int): DoubleArray {
-        val linearR = linearized(redFromArgb(argb))
-        val linearG = linearized(greenFromArgb(argb))
-        val linearB = linearized(blueFromArgb(argb))
-        val matrix = SRGB_TO_XYZ
-        val x = matrix[0][0] * linearR + matrix[0][1] * linearG + matrix[0][2] * linearB
-        val y = matrix[1][0] * linearR + matrix[1][1] * linearG + matrix[1][2] * linearB
-        val z = matrix[2][0] * linearR + matrix[2][1] * linearG + matrix[2][2] * linearB
-        val whitePoint = WHITE_POINT_D65
-        val xNormalized = x / whitePoint[0]
-        val yNormalized = y / whitePoint[1]
-        val zNormalized = z / whitePoint[2]
-        val fx = labF(xNormalized)
-        val fy = labF(yNormalized)
-        val fz = labF(zNormalized)
-        val l = 116.0 * fy - 16
-        val a = 500.0 * (fx - fy)
-        val b = 200.0 * (fy - fz)
-        return doubleArrayOf(l, a, b)
     }
 
     /**
@@ -184,23 +111,6 @@ object ColorUtils {
     }
 
     /**
-     * Converts a Y value to an L* value.
-     *
-     *
-     * L* in L*a*b* and Y in XYZ measure the same quantity, luminance.
-     *
-     *
-     * L* measures perceptual luminance, a linear scale. Y in XYZ measures relative luminance, a
-     * logarithmic scale.
-     *
-     * @param y Y in XYZ
-     * @return L* in L*a*b*
-     */
-    fun lstarFromY(y: Double): Double {
-        return labF(y / 100.0) * 116.0 - 16.0
-    }
-
-    /**
      * Linearizes an RGB component.
      *
      * @param rgbComponent 0 <= rgb_component <= 255, represents R/G/B channel
@@ -223,8 +133,7 @@ object ColorUtils {
      */
     fun delinearized(rgbComponent: Double): Int {
         val normalized = rgbComponent / 100.0
-        var delinearized = 0.0
-        delinearized = if (normalized <= 0.0031308) {
+        val delinearized: Double = if (normalized <= 0.0031308) {
             normalized * 12.92
         } else {
             1.055 * normalized.pow(1.0 / 2.4) - 0.055
