+++ Android/mobile/src/main/java/io/material/hct/material/utils/MathUtils.kt
@@ -14,9 +14,7 @@
  * limitations under the License.
  */
 // This file is automatically generated. Do not modify it.
-package io.material.utils
-
-import kotlin.math.abs
+package io.material.hct.material.utils
 
 /** Utility methods for mathematical operations.  */
 object MathUtils {
@@ -73,20 +71,6 @@ object MathUtils {
     }
 
     /**
-     * Sanitizes a degree measure as an integer.
-     *
-     * @return a degree measure between 0 (inclusive) and 360 (exclusive).
-     */
-    fun sanitizeDegreesInt(degrees: Int): Int {
-        var degrees = degrees
-        degrees %= 360
-        if (degrees < 0) {
-            degrees += 360
-        }
-        return degrees
-    }
-
-    /**
      * Sanitizes a degree measure as a floating-point number.
      *
      * @return a degree measure between 0.0 (inclusive) and 360.0 (exclusive).
@@ -98,28 +82,6 @@ object MathUtils {
             degrees += 360.0
         }
         return degrees
-    }
-
-    /**
-     * Sign of direction change needed to travel from one angle to another.
-     *
-     *
-     * For angles that are 180 degrees apart from each other, both directions have the same travel
-     * distance, so either direction is shortest. The value 1.0 is returned in this case.
-     *
-     * @param from The angle travel starts from, in degrees.
-     * @param to The angle travel ends at, in degrees.
-     * @return -1 if decreasing from leads to the shortest travel distance, 1 if increasing from leads
-     * to the shortest travel distance.
-     */
-    fun rotationDirection(from: Double, to: Double): Double {
-        val increasingDifference = sanitizeDegreesDouble(to - from)
-        return if (increasingDifference <= 180.0) 1.0 else -1.0
-    }
-
-    /** Distance of two points on a circle, represented using degrees.  */
-    fun differenceDegrees(a: Double, b: Double): Double {
-        return 180.0 - abs(abs(a - b) - 180.0)
     }
 
     /** Multiplies a 1x3 row vector with a 3x3 matrix.  */
