+++ Android/mobile/src/main/res/values-en/strings.xml
@@ -121,10 +121,5 @@
     <string name="all_languages">All languages</string>
     <string name="system_settings">System settings</string>
     <string name="predictive_back">Predictive Back Gesture</string>
-    <string name="predictive_back">Predictive Back Gesture</string>
-    <string name="predictive_back">Predictive Back Gesture</string>
-    <string name="predictive_back">Predictive Back Gesture</string>
-    <string name="predictive_back">Predictive Back Gesture</string>
-    <string name="predictive_back">Predictive Back Gesture</string>
     <string name="predictive_back_desc">Enable a smoother back gesture experience</string>
 </resources>
